# swagger/data-exchange-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0",
    "title": "Data Exchange"
  },
  "host": "api.smartcity.kmitl.io",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Get all collection Metadata",
        "parameters": [
          {
            "name": "collectionId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "collectionName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "open",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string",
            "description": "Basic auth or user token for get private collection"
          }
        ],
        "responses": {
          "200": {
            "description": "operation success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Collection"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Create new collection",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Collection_req"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Collection": {
      "type": "object",
      "properties": {
        "collectionId": {
          "type": "string"
        },
        "collectionName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "description": "[\"timeseries\"|\"geotemporal\"|\"keyvalue\"]"
        },
        "category": {
          "type": "string"
        },
        "encryptionLevel": {
          "type": "integer",
          "description": "0 - the data is not encrypted(High optimization and Indexing), 1 - the data is medium encrypted (Medium optimization and Indexing) , 2 - the data is strong encrypted (No optimization and Indexing)",
          "default": 0
        },
        "example": {
          "type": "object"
        },
        "columns": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "column name"
              },
              "type": {
                "type": "string",
                "description": "[\"varchar\"|\"int\"|\"bigint\"|\"double\"|\"text\"|\"timestamp\"]"
              },
              "indexed": {
                "type": "boolean",
                "description": "[true|false]"
              }
            }
          },
          "description": "For 'timeseries' must have column 'ts' type 'timestamp' and for 'geotemporal' must have addition column 'lat' and 'lng' type 'double'"
        },
        "createdAt": {
          "type": "number"
        },
        "isOpen": {
          "type": "boolean"
        }
      }
    },
    "Collection_req": {
      "type": "object",
      "properties": {
        "collectionName": {
          "type": "string"
        },
        "endPoint": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "description": "location of data",
              "example": "local or remote"
            },
            "url": {
              "type": "string",
              "description": "url for request (remote only)"
            },
            "headers": {
              "type": "object",
              "description": "headers for request (remote only)",
              "properties": {
                "headers_0": {
                  "type": "string"
                },
                "headers_1": {
                  "type": "string"
                },
                "header_...": {
                  "type": "string"
                }
              }
            },
            "queryString": {
              "type": "object",
              "description": "query string for request (remote only)",
              "properties": {
                "queryString_1": {
                  "type": "string"
                },
                "queryString_2": {
                  "type": "string"
                },
                "queryString_...": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "description": "[\"timeseries\"|\"geotemporal\"|\"keyvalue\"]"
        },
        "category": {
          "type": "string"
        },
        "encryptionLevel": {
          "type": "integer",
          "description": "0 - the data is not encrypted(High optimization and Indexing), 1 - the data is medium encrypted (Medium optimization and Indexing) , 2 - the data is strong encrypted (No optimization and Indexing)",
          "default": 0
        },
        "example": {
          "type": "object"
        },
        "columns": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "column name"
              },
              "type": {
                "type": "string",
                "description": "[\"varchar\"|\"int\"|\"bigint\"|\"double\"|\"text\"|\"timestamp\"]"
              },
              "indexed": {
                "type": "boolean",
                "description": "[true|false]"
              }
            }
          },
          "description": "For 'timeseries' must have column 'ts' type 'timestamp' and for 'geotemporal' must have addition column 'lat' and 'lng' type 'double'"
        },
        "isOpen": {
          "type": "boolean"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    },
    "userToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "User token"
    },
    "ticket": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Ticket"
    }
  }
}